EOC-Operations/requirements.txt
asgiref==3.4.1
coverage==6.2
Django==3.2.9
django-admin-interface==0.18.3
django-colorfield==0.5.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-geoposition==0.3.0
django-google-maps==0.12.4
django-leaflet==0.28.2
django-location-field==2.1.0
djangorestframework==3.13.1
djangorestframework-gis==0.17
psycopg2==2.9.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
